backend/requirements.txt
fastapi==0.115.12
uvicorn==0.23.2
motor==3.1.2
pymongo==4.4.1
pydantic==2.10.6
pydantic-settings==2.2.1
python-jose==3.3.0
passlib==1.7.4
python-dotenv==1.0.0
bcrypt==4.0.1
email-validator==2.0.0
python-multipart==0.0.6
cryptography==42.0.8
google-cloud-vision==3.7.2
requests==2.31.0
aiosqlite==0.21.0
anyio==4.9.0
python-dotenv
pytesseract==0.3.10
PyPDF2==3.0.1
<<<<<<< HEAD
pdfplumber==0.10.4
nltk==3.9.1

